Refactor utils/libraries + fix tests
