+++ src/components/auth/ForgetPasswordForm.tsx
@@ -3,7 +3,7 @@ import { Button, Field, Heading, Input, VStack } from '@chakra-ui/react';
 import React, { useState } from 'react';
 import { useForm } from 'react-hook-form';
 import { ForgetPasswordFormValues } from '@/types/auth';
-import { supabase } from '@/utils/supabase';
+import { supabase } from '@/lib/supabase';
 
 export default function ForgetPasswordForm() {
   const [isLoading, setIsLoading] = useState<boolean>(false);
